Some kind of a refactoring

+++ src/Controllers/CartController.cpp
@@ -1,35 +1,31 @@
 #include <Controllers/CartController.hpp>
-#include <Database/DatabaseManager.hpp>
-#include <Models/Product.hpp>
-#include <Models/User.hpp>
-#include <Models/Cart.hpp>
 
-namespace models = drogon_model::arthobby;
+#include <Models/Cart.hpp>
+#include <Models/Product.hpp>
+
+namespace Models = drogon_model::arthobby;
+
+namespace Controllers
+{
 
 void CartController::getCart(const HttpRequestPtr& req, Callback&& callback)
 {
     if(req->getParameter("user_id").empty())
     {
-        auto response = HttpResponse::newHttpResponse();
-        response->setStatusCode(drogon::k401Unauthorized);
+        const auto response = HttpResponse::newHttpResponse();
+        response->setStatusCode(k401Unauthorized);
 
         callback(response);
 
         return;
     }
 
-    static auto dbClient = DatabaseManager::get().getDbClient();
-    static auto& mapper = DatabaseManager::get().getMapper<models::Cart>();
+    static auto dbClient = app().getDbClient();
+    static auto mapper = orm::Mapper<Models::Cart>(dbClient);
 
     int userId = std::stoi(req->getParameter("user_id"));
 
-    auto cart = mapper.findBy(
-        orm::Criteria(
-            "user_id",
-            orm::CompareOperator::EQ,
-            userId
-        )
-    );
+    const auto cart = mapper.findBy({ "user_id", userId });
 
     Json::Value json;
     int counter = 0;
@@ -39,7 +35,7 @@ void CartController::getCart(const HttpRequestPtr& req, Callback&& callback)
         json[counter++]["quantity"] = *i.getQuantity();
     }
 
-    auto response = HttpResponse::newHttpJsonResponse(json);
+    const auto response = HttpResponse::newHttpJsonResponse(json);
 
     callback(response);
 }
@@ -48,36 +44,34 @@ void CartController::addToCart(const HttpRequestPtr& req, Callback&& callback)
 {
     if(req->getParameter("user_id").empty())
     {
-        auto response = HttpResponse::newHttpResponse();
-        response->setStatusCode(drogon::k401Unauthorized);
+        const auto response = HttpResponse::newHttpResponse();
+        response->setStatusCode(k401Unauthorized);
 
         callback(response);
 
         return;
     }
 
-    static auto dbClient = DatabaseManager::get().getDbClient();
-    static auto& mapper = DatabaseManager::get().getMapper<models::Cart>();
+    static auto dbClient = app().getDbClient();
+    static auto mapper = orm::Mapper<Models::Cart>(dbClient);
 
     int userId = std::stoi(req->getParameter("user_id"));
     int productId = (*req->getJsonObject())["product_id"].asInt();
 
-    // If we already have such cart entry, just update the quantity
+    // If we already have such cart entry, update the quantity
     try
     {
         auto cart = mapper.findOne(
-            orm::Criteria(
-                orm::CustomSql("user_id = $? AND product_id = $?"),
-                userId, productId
-            )
+            orm::Criteria("user_id", userId)
+            && orm::Criteria("product_id", productId)
         );
 
-        cart.setQuantity(*cart.getQuantity().get() + 1);
+        cart.setQuantity(*cart.getQuantity() + 1);
         mapper.update(cart);
     }
     catch(...) // Otherwise, we will create a new one
     {
-        models::Cart cart;
+        Models::Cart cart;
         cart.setUserId(userId);
         cart.setProductId(productId);
         cart.setQuantity(1);
@@ -92,16 +86,16 @@ void CartController::removeFromCart(const HttpRequestPtr& req, Callback&& callba
 {
     if(req->getParameter("user_id").empty())
     {
-        auto response = HttpResponse::newHttpResponse();
-        response->setStatusCode(drogon::k401Unauthorized);
+        const auto response = HttpResponse::newHttpResponse();
+        response->setStatusCode(k401Unauthorized);
 
         callback(response);
 
         return;
     }
 
-    static auto dbClient = DatabaseManager::get().getDbClient();
-    static auto& mapper = DatabaseManager::get().getMapper<models::Cart>();
+    static auto dbClient = app().getDbClient();
+    static auto mapper = orm::Mapper<Models::Cart>(dbClient);
 
     int userId = std::stoi(req->getParameter("user_id"));
     int productId = (*req->getJsonObject())["product_id"].asInt();
@@ -109,15 +103,13 @@ void CartController::removeFromCart(const HttpRequestPtr& req, Callback&& callba
     try
     {
         auto cart = mapper.findOne(
-            orm::Criteria(
-                orm::CustomSql("user_id = $? AND product_id = $?"),
-                userId, productId
-            )
+            orm::Criteria("user_id", userId)
+            && orm::Criteria("product_id", productId)
         );
 
-        if(*cart.getQuantity().get() > 1)
+        if(*cart.getQuantity() > 1)
         {
-            cart.setQuantity(*cart.getQuantity().get() - 1);
+            cart.setQuantity(*cart.getQuantity() - 1);
             mapper.update(cart);
         }
         else
@@ -125,7 +117,7 @@ void CartController::removeFromCart(const HttpRequestPtr& req, Callback&& callba
     }
     catch(...)
     {
-        // Nothing to delete
+        // Something we can ignore
     }
 
     callback(HttpResponse::newHttpResponse());
@@ -135,27 +127,29 @@ void CartController::clearCart(const HttpRequestPtr& req, Callback&& callback)
 {
     if(req->getParameter("user_id").empty())
     {
-        auto response = HttpResponse::newHttpResponse();
-        response->setStatusCode(drogon::k401Unauthorized);
+        const auto response = HttpResponse::newHttpResponse();
+        response->setStatusCode(k401Unauthorized);
 
         callback(response);
 
         return;
     }
 
-    static auto dbClient = DatabaseManager::get().getDbClient();
-    static auto& mapper = DatabaseManager::get().getMapper<models::Cart>();
+    static auto dbClient = app().getDbClient();
+    static auto mapper = orm::Mapper<Models::Cart>(dbClient);
 
     int userId = std::stoi(req->getParameter("user_id"));
 
-    // May throw an exception, but I won't pay attention!!!
-    mapper.deleteBy(
-        orm::Criteria(
-            "user_id",
-            orm::CompareOperator::EQ,
-            userId
-        )
-    );
+    try
+    {
+        mapper.deleteBy({ "user_id", userId });
+    }
+    catch(...)
+    {
+        // No need to handle this I guess
+    }
 
     callback(HttpResponse::newHttpResponse());
+}
+
 }
